Fix flaky capture_produces_frames: isolate SIGSEGV-handler-replacement test
`capture_debug_asserts_when_handler_replaced` replaced the *process-wide*
SIGSEGV handler with SIG_IGN to verify `Unwinder::capture`'s `debug_assert!`
fires. Its comment assumed nextest's process-per-test isolation, but CI runs
plain `cargo test`, whose default harness runs all unit tests on threads in a
single process. During the replacement window, sibling test threads (notably
`capture_produces_frames`) called `capture()` and hit
`debug_assert!(verify_handler())`, panicking intermittently.

Move the test into its own integration-test binary (tests/handler_replacement.rs).
Each file under tests/ compiles to a separate binary and runs in its own
process, so the global handler mutation can no longer race with the unit tests.
